requirements.txt
# AI INTELLIGENCE ZONE — Control Arena
# Production Dependencies

# Core Framework
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache & Sessions
redis==5.0.1
Flask-Session==0.5.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1

# Validation
jsonschema==4.21.1
marshmallow==3.20.2

# Real-Time
eventlet==0.35.1
python-socketio==5.11.1

# Task Queue
celery==5.3.6

# Security
bleach==6.1.0
cryptography==42.0.2

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[flask]==1.40.4
prometheus-flask-instrumentator==7.0.0

# Utilities
python-dateutil==2.8.2
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Testing
pytest==8.0.1
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
